feat(sap-aicore-llm)!: extract SapCoreAIProvider into @mcp-abap-adt/sap-aicore-llm
Moves SapCoreAIProvider and related types from llm-agent-server into a new standalone package
with @sap-ai-sdk/orchestration as a direct dependency. Reduces server dependency bloat and
enables independent release cycle for SAP-specific integrations.

- Create packages/sap-aicore-llm with SapCoreAIProvider, SapCoreAIConfig, SapAICoreCredentials
- git mv sap-core-ai.ts → sap-core-ai-provider.ts (file path consistency)
- Update server's llm-providers barrel to re-export from @mcp-abap-adt/sap-aicore-llm
- Update all server imports to use package import instead of relative imports
- Add @types/node to root devDependencies for type support
- All 16 sap-aicore-llm tests pass, all 481 server tests pass
