+++ package.json
@@ -18,6 +18,7 @@
   "devDependencies": {
     "@biomejs/biome": "^2.4.9",
     "@changesets/cli": "^2.31.0",
+    "@types/node": "^22.0.0",
     "typescript": "^6.0.2",
     "tsx": "^4.21.0",
     "rimraf": "^5.0.0"
