+++ packages/llm-agent-server/src/llm-providers/index.ts
@@ -14,5 +14,8 @@ export { BaseLLMProvider, type LLMProvider } from '@mcp-abap-adt/llm-agent';
 // Legacy exports (deprecated - use SapCoreAIProvider instead)
 // These are kept for backward compatibility but will be removed in future versions
 export { type OpenAIConfig, OpenAIProvider } from '@mcp-abap-adt/openai-llm';
+export {
+  type SapCoreAIConfig,
+  SapCoreAIProvider,
+} from '@mcp-abap-adt/sap-aicore-llm';
 export { type DeepSeekConfig, DeepSeekProvider } from './deepseek.js';
-export { type SapCoreAIConfig, SapCoreAIProvider } from './sap-core-ai.js';
