+++ packages/sap-aicore-llm/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@mcp-abap-adt/sap-aicore-llm",
+  "version": "11.0.0",
+  "description": "SAP AI Core LLM provider (ILlm) for @mcp-abap-adt/llm-agent.",
+  "type": "module",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist",
+    "README.md",
+    "LICENSE"
+  ],
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "clean": "tsc -p tsconfig.json --clean",
+    "test": "node --import tsx/esm --test --test-reporter=spec 'src/**/*.test.ts'"
+  },
+  "dependencies": {
+    "@mcp-abap-adt/llm-agent": "*",
+    "@sap-ai-sdk/ai-api": "^2.9.0",
+    "@sap-ai-sdk/orchestration": "^2.9.0"
+  },
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fr0ster/llm-agent.git"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
