+++ packages/sap-aicore-llm/src/__tests__/sap-core-ai-provider.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 import type { Message } from '@mcp-abap-adt/llm-agent';
-import { SapCoreAIProvider } from '../sap-core-ai.js';
+import { SapCoreAIProvider } from '../sap-core-ai-provider.js';
 
 // ---------------------------------------------------------------------------
 // Constructor
