+++ packages/sap-aicore-llm/src/index.ts
@@ -1,0 +1,5 @@
+export type {
+  SapAICoreCredentials,
+  SapCoreAIConfig,
+} from './sap-core-ai-provider.js';
+export { SapCoreAIProvider } from './sap-core-ai-provider.js';
